requirements.txt
langchain  
google-generativeai
faiss-cpu
langchain_community 
langchain-google-genai 
pypdf 
PyPDF2
streamlit
python-dotenv